Backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
tensorflow
opencv-python
opencv-contrib-python
numpy
pandas
matplotlib
motor
pydantic-settings
pymongo
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
fastapi-mail
mtcnn
